fix [#85] 멤버 목록 레이아웃 수정 - 할 일 목록과 동일하게 top align 맞춤
